fix: make filters consistent
