+++ openmetadata-ui/src/main/resources/ui/src/components/Assets/AssetsSelectionModal/asset-selection-model.style.less
@@ -14,7 +14,7 @@
 @import url('../../../styles/variables.less');
 
 .asset-selection-model-card.table-data-card-container {
-  box-shadow: none;
+  box-shadow: none !important;
 }
 .asset-selection-model-card.table-data-card-container:hover {
   box-shadow: 2px 3px 5px rgba(0, 0, 0, 0.13);
